refactor: auth 인증 state를 state와 action 으로 분리하여 store로 통합 #37
